refactor: AuthServiceUtil 을 사용하도록 변경 - #135
