+++ src/main/java/org/mju_likelion/festival/auth/service/AuthQueryService.java
@@ -1,34 +1,26 @@
 package org.mju_likelion.festival.auth.service;
 
-import lombok.AllArgsConstructor;
+import lombok.RequiredArgsConstructor;
 import org.mju_likelion.festival.admin.domain.Admin;
-import org.mju_likelion.festival.admin.domain.repository.AdminJpaRepository;
-import org.mju_likelion.festival.auth.domain.RsaKey;
-import org.mju_likelion.festival.auth.domain.RsaKeyStrategy;
 import org.mju_likelion.festival.auth.dto.request.AdminLoginRequest;
 import org.mju_likelion.festival.auth.dto.response.AdminLoginResponse;
 import org.mju_likelion.festival.auth.dto.response.KeyResponse;
 import org.mju_likelion.festival.auth.util.jwt.AuthenticationRole;
-import org.mju_likelion.festival.auth.util.jwt.JwtUtil;
-import org.mju_likelion.festival.auth.util.jwt.Payload;
+import org.mju_likelion.festival.auth.util.key.RsaKey;
+import org.mju_likelion.festival.auth.util.key.RsaKeyStrategy;
 import org.mju_likelion.festival.auth.util.key.manager.RsaKeyManager;
-import org.mju_likelion.festival.auth.util.key.manager.RsaKeyManagerContext;
-import org.mju_likelion.festival.common.exception.UnauthorizedException;
-import org.mju_likelion.festival.common.exception.type.ErrorType;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
-@AllArgsConstructor
+@RequiredArgsConstructor
 @Transactional(readOnly = true)
 public class AuthQueryService {
 
-  private final RsaKeyManagerContext rsaKeyManagerContext;
-  private final AdminJpaRepository adminJpaRepository;
-  private final JwtUtil jwtUtil;
+  private final AuthServiceUtil authServiceUtil;
 
   public KeyResponse getKey() {
-    RsaKeyManager rsaKeyManager = rsaKeyManagerContext.rsaKeyManager();
+    RsaKeyManager rsaKeyManager = authServiceUtil.getRsaKeyManager();
 
     RsaKey rsaKey = rsaKeyManager.generateRsaKey();
 
@@ -43,30 +35,19 @@ public class AuthQueryService {
       final AdminLoginRequest adminLoginRequest,
       final RsaKeyStrategy rsaKeyStrategy) {
 
-    RsaKeyManager rsaKeyManager = rsaKeyManagerContext.rsaKeyManager(rsaKeyStrategy);
+    RsaKeyManager rsaKeyManager = authServiceUtil.getRsaKeyManager(rsaKeyStrategy);
 
     String key = adminLoginRequest.getKey();
     String loginId = rsaKeyManager.decryptByKey(adminLoginRequest.getEncryptedLoginId(), key);
     String password = rsaKeyManager.decryptByKey(adminLoginRequest.getEncryptedPassword(), key);
 
-    Admin admin = getExistingAdmin(loginId, password);
+    Admin admin = authServiceUtil.getExistingAdmin(loginId, password);
 
-    String accessToken = jwtUtil.create(
-        new Payload(admin.getId(), AuthenticationRole.from(admin.getRole()))
-    );
+    String accessToken = authServiceUtil.createAccessToken(admin.getId(),
+        AuthenticationRole.from(admin.getRole()));
 
     return AdminLoginResponse.of(accessToken, admin.getRole());
   }
 
-  /**
-   * 관리자가 존재하는지 확인하고, 존재하는 경우 사용자를 반환한다.
-   *
-   * @param loginId  로그인 ID
-   * @param password 비밀번호
-   * @return 관리자
-   */
-  private Admin getExistingAdmin(final String loginId, final String password) {
-    return adminJpaRepository.findByLoginIdAndPassword(loginId, password)
-        .orElseThrow(() -> new UnauthorizedException(ErrorType.INVALID_CREDENTIALS_ERROR));
-  }
+
 }
