+++ src/main/java/org/mju_likelion/festival/auth/service/AuthService.java
@@ -6,19 +6,14 @@ import java.util.Optional;
 import java.util.Set;
 import java.util.UUID;
 import java.util.stream.Collectors;
-import lombok.AllArgsConstructor;
-import org.mju_likelion.festival.auth.domain.RsaKeyStrategy;
+import lombok.RequiredArgsConstructor;
 import org.mju_likelion.festival.auth.dto.request.UserLoginRequest;
 import org.mju_likelion.festival.auth.dto.response.UserLoginResponse;
 import org.mju_likelion.festival.auth.util.jwt.AuthenticationRole;
-import org.mju_likelion.festival.auth.util.jwt.JwtUtil;
-import org.mju_likelion.festival.auth.util.jwt.Payload;
+import org.mju_likelion.festival.auth.util.key.RsaKeyStrategy;
 import org.mju_likelion.festival.auth.util.key.manager.RsaKeyManager;
-import org.mju_likelion.festival.auth.util.key.manager.RsaKeyManagerContext;
 import org.mju_likelion.festival.common.exception.BadRequestException;
-import org.mju_likelion.festival.common.exception.UnauthorizedException;
 import org.mju_likelion.festival.common.exception.type.ErrorType;
-import org.mju_likelion.festival.common.util.api.MjuApiUtil;
 import org.mju_likelion.festival.term.domain.Term;
 import org.mju_likelion.festival.term.domain.repository.TermJpaRepository;
 import org.mju_likelion.festival.user.domain.User;
@@ -27,31 +22,32 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
-@AllArgsConstructor
+@RequiredArgsConstructor
+@Transactional
 public class AuthService {
 
-  private final RsaKeyManagerContext rsaKeyManagerContext;
-  private final MjuApiUtil mjuApiUtil;
+  private final AuthServiceUtil authServiceUtil;
   private final UserJpaRepository userJpaRepository;
   private final TermJpaRepository termJpaRepository;
-  private final JwtUtil jwtUtil;
 
-  @Transactional
+
   public UserLoginResponse userLogin(
       final UserLoginRequest userLoginRequest,
       final RsaKeyStrategy rsaKeyStrategy) {
 
-    RsaKeyManager rsaKeyManager = rsaKeyManagerContext.rsaKeyManager(rsaKeyStrategy);
+    RsaKeyManager rsaKeyManager = authServiceUtil.getRsaKeyManager(rsaKeyStrategy);
 
     String key = userLoginRequest.getKey();
     String studentId = rsaKeyManager.decryptByKey(userLoginRequest.getEncryptedStudentId(), key);
     String password = rsaKeyManager.decryptByKey(userLoginRequest.getEncryptedPassword(), key);
 
-    validateUser(studentId, password);
+    authServiceUtil.validateUser(studentId, password);
 
-    UUID userId = saveOrGetUserId(studentId, userLoginRequest.getTerms());
+    List<Term> terms = getValidTerms(userLoginRequest.getTerms());
 
-    String accessToken = jwtUtil.create(new Payload(userId, AuthenticationRole.USER));
+    UUID userId = saveOrGetUserId(studentId, terms);
+
+    String accessToken = authServiceUtil.createAccessToken(userId, AuthenticationRole.USER);
     return UserLoginResponse.of(accessToken);
   }
 
@@ -64,12 +60,10 @@ public class AuthService {
    * @param terms     사용자가 동의한 약관
    * @return 사용자의 ID
    */
-  private UUID saveOrGetUserId(final String studentId, final Map<UUID, Boolean> terms) {
-    List<Term> validTerms = getValidTerms(terms);
-
+  private UUID saveOrGetUserId(final String studentId, final List<Term> terms) {
     Optional<UUID> userId = userJpaRepository.findIdByStudentId(studentId);
 
-    return userId.orElseGet(() -> saveUser(studentId, validTerms));
+    return userId.orElseGet(() -> saveUser(studentId, terms));
   }
 
   /**
@@ -111,18 +105,6 @@ public class AuthService {
     Set<UUID> validTermIds = termsInDb.stream().map(Term::getId).collect(Collectors.toSet());
     if (!termIds.containsAll(validTermIds)) {
       throw new BadRequestException(ErrorType.MISSING_TERM_ERROR);
-    }
-  }
-
-  /**
-   * 사용자의 학번과 비밀번호를 검증한다. 명지대학교 MSI API 를 이용하여 사용자의 존재 여부를 확인한다.
-   *
-   * @param studentId 학번
-   * @param password  비밀번호
-   */
-  private void validateUser(final String studentId, final String password) {
-    if (!mjuApiUtil.doUserCheck(studentId, password)) {
-      throw new UnauthorizedException(ErrorType.INVALID_CREDENTIALS_ERROR);
     }
   }
 }
